+++ src/App.tsx
@@ -3,9 +3,11 @@ import AppLoading from 'expo-app-loading'
 import { createStackNavigator } from '@react-navigation/stack'
 import { useFonts, Play_400Regular } from '@expo-google-fonts/play'
 import { Padauk_700Bold } from '@expo-google-fonts/padauk'
+import { Mukta_400Regular, Mukta_700Bold } from '@expo-google-fonts/mukta'
 
 import SignIn from './screens/SignIn'
 import SignInEmail from './screens/SignInEmail'
+import SelectCourses from './screens/SelectCourses'
 
 const Stack = createStackNavigator()
 
@@ -13,7 +15,9 @@ const App = () => {
   const [fondLoaded] = useFonts({
     Catamaran: require('../assets/fonts/Catamaran-VariableFont_wght.ttf'),
     Padauk_700Bold,
-    Play_400Regular
+    Play_400Regular,
+    Mukta_400Regular,
+    Mukta_700Bold
   })
 
   if (!fondLoaded)
@@ -25,8 +29,9 @@ const App = () => {
         headerShown: false
       }}
     >
-      <Stack.Screen name='SignIn' component={SignIn}/>
-      <Stack.Screen name='SignInEmail' component={SignInEmail}/>
+      <Stack.Screen name='SignIn' component={SignIn} />
+      <Stack.Screen name='SignInEmail' component={SignInEmail} />
+      <Stack.Screen name='SelectCourses' component={SelectCourses} />
     </Stack.Navigator>
   )
 }