+++ src/utils/index.ts
@@ -1,3 +1,49 @@
 import COLORS from './colors'
 
-export { COLORS }
+interface ICourseData {
+  course  : string
+  icon    : string
+  iconType: string
+  id      : string
+}
+
+const COURSES: ICourseData[] = [
+  {
+    course  : 'Integral Calculus',
+    icon    : 'math-integral',
+    iconType: 'material-community',
+    id      : '1'
+  },
+  {
+    course  : 'Electrostatics',
+    icon    : 'flash',
+    iconType: 'material-community',
+    id      : '2'
+  },
+  {
+    course  : 'Java Programming',
+    icon    : 'language-java',
+    iconType: 'material-community',
+    id      : '3'
+  },
+  {
+    course  : 'Organic Chemistry',
+    icon    : 'chemistry',
+    iconType: 'simple-line-icon',
+    id      : '4'
+  },
+  {
+    course  : 'Statistics',
+    icon    : 'bar-chart',
+    iconType: 'material',
+    id      : '5'
+  },
+  {
+    course  : 'Human-Computer Interaction',
+    icon    : 'computer',
+    iconType: 'material',
+    id      : '6'
+  }
+]
+
+export { COLORS, COURSES }